+++ pkg/platform/service_enable.go
@@ -42,7 +42,7 @@ func serviceEnable() (*cobra.Command, error) {
 	cmd.Use = "enable [flags] deployment name chart"
 	cmd.Short = "Manages Service Installation/Management"
 
-	if err := cli.RegisterFlags(&cmd, flagPlatformStage, flagPlatformEndpoint, flagPlatformName, flagValues); err != nil {
+	if err := cli.RegisterFlags(&cmd, flagPlatformEndpoint, flagPlatformName, flagValues); err != nil {
 		return nil, err
 	}