+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - (Feature) (Platform) Reorganize Envoy Auth
 - (Feature) Ensure Group Service Type
 - (Maintenance) Fix Helm & JWT CVE's
+- (Feature) (Platform) Improve CLI Values
 
 ## [1.2.48](https://github.com/arangodb/kube-arangodb/tree/1.2.48) (2025-05-08)
 - (Maintenance) Extend Documentation